third_party/blink/web_tests/external/wpt/websockets/interfaces/CloseEvent/clean-close_wpt_flags=h2-expected.txt

This is a testharness.js-based test.
FAIL WebSockets: wasClean, true assert_equals: expected true but got false
Harness: the test ran to completion.
